+++ Changelog.md
@@ -34,6 +34,7 @@
 * Add rake task to send a mail to all users [#5111](https://github.com/diaspora/diaspora/pull/5111)
 * Expose which services are configured in /statistics.json [#5121](https://github.com/diaspora/diaspora/pull/5121)
 * In filtered notification views, replace "Mark all as read" with "Mark shown as read" [#5122](https://github.com/diaspora/diaspora/pull/5122)
+* When ignoring a user remove his posts from the stream instantly [#5127](https://github.com/diaspora/diaspora/pull/5127)
 
 # 0.4.0.1